refact: ディレクトリ構成を変更
