+++ api/internal/article/tag.go
@@ -1,4 +1,4 @@
-package tags
+package article
 
 type Tag struct {
 	ID   int64
