--- api/internal/get_auth_me/model.go
@@ -1,10 +1,0 @@
-package me
-
-import "time"
-
-type User struct {
-	ID        int64
-	Name      string
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
