--- api/internal/get_tags/repository.go
@@ -1,49 +1,0 @@
-package tags
-
-import (
-	"context"
-	"database/sql"
-	"errors"
-)
-
-type Repository struct {
-	db *sql.DB
-}
-
-func NewRepository(db *sql.DB) *Repository {
-	return &Repository{db: db}
-}
-
-func (r *Repository) List(ctx context.Context) ([]Tag, error) {
-	if r == nil || r.db == nil {
-		return nil, errors.New("tags repository is not configured")
-	}
-
-	const query = `
-		SELECT id, name
-		FROM tags
-		ORDER BY name ASC, id ASC
-	`
-
-	rows, err := r.db.QueryContext(ctx, query)
-	if err != nil {
-		return nil, err
-	}
-	defer rows.Close()
-
-	tags := make([]Tag, 0)
-	for rows.Next() {
-		var tag Tag
-		if err := rows.Scan(&tag.ID, &tag.Name); err != nil {
-			return nil, err
-		}
-
-		tags = append(tags, tag)
-	}
-
-	if err := rows.Err(); err != nil {
-		return nil, err
-	}
-
-	return tags, nil
-}
