--- api/internal/post_article/handler.go
@@ -1,155 +1,0 @@
-package article
-
-import (
-	"context"
-	"errors"
-	"net/http"
-	"strings"
-	"time"
-	"unicode/utf8"
-
-	"github.com/Baymax6s/KOBE-Tech/api/internal/auth"
-	"github.com/gin-gonic/gin"
-)
-
-type createArticleRequest struct {
-	Title    string   `json:"title"`
-	Content  string   `json:"content"`
-	TagNames []string `json:"tags" minLength:"1" maxLength:"10"`
-} // @name server.createArticleRequest
-
-type TagJSON struct {
-	ID   int64  `json:"id" binding:"required"`
-	Name string `json:"name" binding:"required"`
-} // @name server.articleTagJSONResponse
-
-type ArticleJSON struct {
-	ID        int64     `json:"id" binding:"required"`
-	Title     string    `json:"title" binding:"required"`
-	Content   string    `json:"content" binding:"required"`
-	UserID    int64     `json:"user_id" binding:"required"`
-	Tags      []TagJSON `json:"tags" binding:"required"`
-	CreatedAt time.Time `json:"created_at" binding:"required"`
-	UpdatedAt time.Time `json:"updated_at" binding:"required"`
-} // @name server.createArticleResponse
-
-type ErrorResponse struct {
-	Message string `json:"message"`
-} // @name server.articleErrorResponse
-
-type Handler struct {
-	repo *Repository
-}
-
-func NewHandler(repo *Repository) *Handler {
-	return &Handler{repo: repo}
-}
-
-func (h *Handler) RegisterRoutes(router gin.IRouter) {
-	router.POST("/articles", h.createArticleHandler)
-}
-
-// createArticleHandler godoc
-//
-//	@Summary		Create article
-//	@Description	Create article API.
-//	@Tags			articles
-//	@Accept			json
-//	@Produce		json
-//	@Param			request	body		createArticleRequest	true	"Create article request"
-//	@Success		201		{object}	ArticleJSON
-//	@Failure		400		{object}	ErrorResponse
-//	@Failure		401		{object}	ErrorResponse
-//	@Failure		500		{object}	ErrorResponse
-//	@Security		BearerAuth
-//	@Router			/api/articles [post]
-func (h *Handler) createArticleHandler(c *gin.Context) {
-	var req createArticleRequest
-	if err := c.ShouldBindJSON(&req); err != nil {
-		c.JSON(http.StatusBadRequest, ErrorResponse{Message: "invalid request body"})
-		return
-	}
-
-	userID := auth.MustUserID(c)
-
-	response, err := h.CreateArticle(c.Request.Context(), userID, req.Title, req.Content, req.TagNames)
-	if err != nil {
-		switch {
-		case errors.Is(err, errInvalidRequest), errors.Is(err, errInvalidTagName):
-			c.JSON(http.StatusBadRequest, ErrorResponse{Message: err.Error()})
-		default:
-			c.JSON(http.StatusInternalServerError, ErrorResponse{Message: "failed to create article"})
-		}
-		return
-	}
-
-	c.JSON(http.StatusCreated, response)
-}
-
-var errInvalidRequest = errors.New("title and content are required")
-var errInvalidTagName = errors.New("tags must contain tag names between 1 and 10 characters")
-
-const maxTagNameLength = 10
-
-func (h *Handler) CreateArticle(ctx context.Context, userID int64, title, content string, tagNames []string) (ArticleJSON, error) {
-	if h == nil || h.repo == nil {
-		return ArticleJSON{}, errors.New("post article handler is not configured")
-	}
-
-	title = strings.TrimSpace(title)
-	content = strings.TrimSpace(content)
-	if title == "" || content == "" {
-		return ArticleJSON{}, errInvalidRequest
-	}
-
-	normalizedTagNames, err := normalizeTagNames(tagNames)
-	if err != nil {
-		return ArticleJSON{}, err
-	}
-
-	article, err := h.repo.Create(ctx, title, content, userID, normalizedTagNames)
-	if err != nil {
-		return ArticleJSON{}, err
-	}
-
-	tags := make([]TagJSON, 0, len(article.Tags))
-	for _, tag := range article.Tags {
-		tags = append(tags, TagJSON{
-			ID:   tag.ID,
-			Name: tag.Name,
-		})
-	}
-
-	return ArticleJSON{
-		ID:        article.ID,
-		Title:     article.Title,
-		Content:   article.Content,
-		UserID:    article.UserID,
-		Tags:      tags,
-		CreatedAt: article.CreatedAt,
-		UpdatedAt: article.UpdatedAt,
-	}, nil
-}
-
-func normalizeTagNames(tagNames []string) ([]string, error) {
-	normalizedTagNames := make([]string, 0, len(tagNames))
-	seen := make(map[string]struct{}, len(tagNames))
-
-	for _, tagName := range tagNames {
-		normalizedTagName := strings.ToLower(strings.TrimSpace(tagName))
-		if normalizedTagName == "" {
-			return nil, errInvalidTagName
-		}
-		if utf8.RuneCountInString(normalizedTagName) > maxTagNameLength {
-			return nil, errInvalidTagName
-		}
-		if _, ok := seen[normalizedTagName]; ok {
-			continue
-		}
-
-		seen[normalizedTagName] = struct{}{}
-		normalizedTagNames = append(normalizedTagNames, normalizedTagName)
-	}
-
-	return normalizedTagNames, nil
-}
