--- api/internal/post_article/model.go
@@ -1,18 +1,0 @@
-package article
-
-import "time"
-
-type Article struct {
-	ID        int64
-	Title     string
-	Content   string
-	UserID    int64
-	Tags      []Tag
-	CreatedAt time.Time
-	UpdatedAt time.Time
-}
-
-type Tag struct {
-	ID   int64
-	Name string
-}
